requirements.txt
fastapi
uvicorn
playwright
jinja2
python-dotenv
psycopg2-binary